feat: regenerate against the rebuilt AXO branch
Regenerates from `clever-cloud/axo!1768`, which was rebuilt on current `main` rather than rebased
onto it. The previous generation here came from that branch before it caught up with 496 upstream
commits, so it described a spec that no longer exists.

  52 packages, 1018 operations, 864 models   (was 51 / 948 / 803)

The extra operations are the modules upstream added while the branch drifted. They carry restored
identifiers because the migration was replayed rather than merged — a rebase would have left them
with an id derived from their Rust function name, and nothing would have said so.

**321 of the 323 routes the published SDK shares with this document now carry the operation id it
expects.**

## Generator fixes

Two more divergences between the two generators, which derive the same names independently:

  - **Package names.** `generate-services` snake-cases; `generate-builder` only replaced hyphens,
    so a `NetworkGroup` tag produced an import of `services/NetworkGroup` while the directory on
    disk was `services/network_group`. Go package names are lower-case anyway.
  - **Acronym expansion.** The tables were unified in the previous commit, but not the code that
    applies them: the builder did a blanket `ReplaceAll`, turning `Ids` into `IDs`, where services
    replaces only at the end of a word or before another capital. One operation came out as
    `PostgresqlInternalIDsByOwner` on one side and `PostgresqlInternalIdsByOwner` on the other.
    Unifying the data was not enough — the algorithm had to match too.

That makes four such divergences found by generating: the acronym table, the package-name
exceptions, the package-name normalisation, and now the expansion algorithm. Every one is the same
shape — a rule implemented twice with nothing keeping the copies honest, which is exactly the
defect this whole effort addresses on the spec side.

## What generating found in the spec, fixed upstream

  - **45 loadbalancer operations declared a path template without declaring its parameters.** Six
    of them broke generation; the rest were simply invisible to any generator or documentation UI.
  - `/v4/products/zones` declared `tag` and `ownerId` as *path* parameters — `IntoParams` defaults
    to `Path` when `parameter_in` is omitted — so a generated client would have formatted query
    values into the URL.
  - The PostgreSQL privilege routes published `objectId` three times in one path, which OpenAPI
    forbids; the SDK's function for that route emits `%!s(MISSING)` today and always has.

`go build ./...` and `go vet ./...` are clean.

+++ models/keycloak_consumption_detail.go
@@ -2,8 +2,8 @@
 
 package models
 
-// KeycloakConsumptionDetail Consumption detail with interval granularity.  Source: references/legacy/ovd/modules/keycloak/src...
+// KeycloakConsumptionDetail Per-interval breakdown of a consumption item.  Source: references/legacy/ovd/core/src/main/scala/...
 type KeycloakConsumptionDetail struct {
-	Interval   string `json:"interval"`   // Interval type (e.g. "daily")
-	Quantities []int  `json:"quantities"` // Per-interval quantities (filled with 1s for keycloak licence)
+	Interval   string    `json:"interval"`   // ISO-8601 period. `Interval.daily` encodes via `toISO8601` to **"P1D"** — the Scala symbol name `d...
+	Quantities []float64 `json:"quantities"` // One `1.0` per active day.
 }
